Reorganize location of new codegen files
