+++ source/MaterialXRuntime/Codegen/RtCodegenImpl.cpp
@@ -3,7 +3,7 @@
 // All rights reserved.  See LICENSE.txt for license.
 //
 
-#include <MaterialXGenShader/Impl/RtCodegenImpl.h>
+#include <MaterialXRuntime/Codegen/RtCodegenImpl.h>
 
 #include <MaterialXRuntime/Private/PvtPath.h>
 #include <MaterialXRuntime/Private/PvtPrim.h>
