+++ source/MaterialXRuntime/Codegen/RtSourceCodeImpl.h
@@ -9,7 +9,7 @@
 /// @file RtSourceCodeImpl.h
 /// TODO: Docs
 
-#include <MaterialXGenShader/Impl/RtCodegenImpl.h>
+#include <MaterialXRuntime/Codegen/RtCodegenImpl.h>
 
 namespace MaterialX
 {
